feature(system-tests): switch from nightwatch to playwright
