+++ System-Tests/run_all_tests.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+for test in test_*
+do
+    node tests.mjs $test
+done
