+++ System-Tests/tests.mjs
@@ -1,0 +1,49 @@
+import process from "process";
+import { firefox, chromium } from "playwright";
+
+const testRunner = async (testPaths) => {
+    const browsers = {
+        firefox: await firefox.launch({ headless: true }),
+        chromium: await chromium.launch({ headless: true })
+    };
+    let testSuites = [];
+    try {
+        testSuites = [
+            // Promise.all resolves as an array with 1 element somehow
+            await Promise.all(testPaths.map(async test => {
+                return { name: test, module: await import(`./${test}`) };
+            }))
+        ][0].map(suite => {
+            return { name: suite.name, module: suite.module.default };
+        });
+    } catch (error) {
+        console.log(error);
+
+    }
+    for (const key of Object.keys(browsers)) {
+        for (const suite of testSuites) {
+            try {
+                let browserContext = await browsers[key].newContext({ acceptDownloads: true });
+                await suite.module(browserContext);
+
+                console.log(`Test: ${suite.name} in browser ${key} ran successfully`);
+                browsers[key].close();
+            } catch (error) {
+                browsers[key].close();
+                console.log(`Test: ${suite.name} in browser ${key} failed`);
+                throw "Tests failed";
+            }
+        }
+    }
+};
+
+const main = async () => {
+    try {
+        const testPaths = process.argv.slice(2);
+        await testRunner(testPaths);
+    } catch (error) {
+        console.error(error);
+        process.exit(1);
+    }
+};
+main();
