--- System-Tests/tests/nightwatch/dashboard/dashboard_device_widget.spec.js
@@ -1,119 +1,0 @@
-const fs = require("fs").promises
-const assert = require('assert')
-const util = require('util');
-const glob = require('glob');
-const setTimeoutAsync = util.promisify(setTimeout);
-const { setServer } = require("../helpers/dataFillers");
-
-module.exports = {
-    "Add server": async function (browser) {
-        setServer(browser)
-    },
-    "Add new device": async function (browser) {
-        await browser.click("#add-device button");
-
-        await browser.setValue(".device:first-child input[name='device-ipv4']", 2);
-        await browser.setValue(".device:first-child input[name='device-name']", "test1");
-        await browser.setValue(".device:first-child select.device-type", "Linux");
-        await browser.pause(2000) // REFACTOR: this is done to wait for the debounce on the inputs
-        browser.expect.elements(".table-row.device").count.to.equal(1);
-    },
-    "Add complex device": async function (browser) {
-        await browser.click("#add-device button");
-
-        await browser.setValue(".device:last-child input[name='device-ipv4']", 3);
-        await browser.setValue(".device:last-child input[name='device-name']", "test2");
-        await browser.setValue(".device:last-child select.device-type", "Linux");
-        await browser.setValue(".device:last-child input[name='MTU']", "1320");
-        await browser.setValue(".device:last-child input[name='additionalDNSServers']", "1.1.1.1,2.2.2.2");
-        await browser.click(".device:last-child input[name='routed']");
-        await browser.pause(2000) // REFACTOR: this is done to wait for the debounce on the inputs
-        browser.expect.elements(".table-row.device").count.to.equal(2);
-    },
-    "Download and verify server configuration": async function (browser) {
-        const downloadLocation = '/tmp/WirtTestDownloads/server.conf'
-        await browser.click("#server-widget #download");
-
-        await setTimeoutAsync(1000, 'waitForDownload').then(async () => {
-            try {
-                const data = await fs.readFile(downloadLocation, "utf8")
-                assert.match(data, /.*ListenPort = 1233.*/)
-
-                await fs.unlink(downloadLocation)
-            } catch (error) {
-                await fs.unlink(downloadLocation)
-                throw new Error(error);
-            }
-
-        })
-    },
-    "Download and verify complex device configuration": async function (browser) {
-        const downloadLocation = '/tmp/WirtTestDownloads/test2.conf'
-        await browser.click(".table-row.device:last-child .download");
-        await setTimeoutAsync(1000, 'waitForDownload').then(async () => {
-            try {
-                const data = await fs.readFile(downloadLocation, "utf8")
-                assert.match(data, /.*DNS = 10\.10\.0\.1,1\.1\.1\.1,2\.2\.2\.2.*/)
-                assert.match(data, /.*MTU = 1320.*/)
-
-                await fs.unlink(downloadLocation)
-            } catch (error) {
-                await fs.unlink(downloadLocation)
-                throw new Error(error);
-            }
-        })
-    },
-    "Download and verify that backup is a simple JSON string": async function (browser) {
-        await browser.click("#export button");
-        await setTimeoutAsync(1000, 'waitForDownload').then(async () => {
-            try {
-                let files = glob.sync("/tmp/WirtTestDownloads/dasboard-backup-*.json");
-                let file = files[files.length - 1];
-                try {
-                    const data = await fs.readFile(file, "utf8")
-                    let json = JSON.parse(data);
-                    assert(typeof json.deviceTypes === 'object')
-                    fs.unlink(file)
-                } catch (error) {
-                    fs.unlink(file)
-                    throw new Error(error);
-                }
-            } catch (error) {
-                throw new Error(error);
-            }
-        })
-    },
-    "MTU is bounded": async function (browser) {
-        await browser.click("#add-device button");
-
-        await browser.setValue("input[name='device-name']", "mtu-test");
-        await browser.setValue("input[name='MTU']", "20000");
-        browser.expect("input[name='MTU']").to.have.property('validity').to.have.property('valid').equals('false');
-        await browser.pause(3000)
-        await browser.click("button.stop");
-    },
-    "additional DNS Servers must be a comma separated list of valid IPv4s": async function (browser) {
-        await browser.click("#add-device button");
-        await browser.setValue("input[name='additionalDNSServers']", "1.1.1.1,2.2.2.t");
-        browser.expect("input[name='additionalDNSServers']").to.have.property('validity').to.have.property('valid').equals('false');
-
-        await browser.setValue("input[name='additionalDNSServers']", "1.1.1.1,2.2.2.");
-        browser.expect("input[name='additionalDNSServers']").to.have.property('validity').to.have.property('valid').equals('false');
-        await browser.pause(3000)
-
-        await browser.click("button.stop");
-    },
-    "correctly deletes device from config": async function (browser) {
-        await browser.click(".table-row.device:first-child .delete");
-        await browser.pause(3000)
-        await setTimeoutAsync(1000, 'waitForUpdate').then(async () => {
-            try {
-                const data = await fs.readFile('/etc/wireguard/server.conf', "utf8")
-                assert.doesNotMatch(data, /.*AllowedIPs= 10\.10\.0\.2*/)
-            } catch (error) {
-                throw new Error(error);
-            }
-        })
-
-    },
-};
