--- System-Tests/tests/nightwatch/dashboard/dashboard_functionality_mobile.spec.js
@@ -1,8 +1,0 @@
-const functionality_tests = require("./dashboard_functionality.spec")
-
-module.exports = {
-    "Resize window": async function (browser) {
-        await browser.resizeWindow(480, 800)
-    },
-    ...functionality_tests
-}
