--- System-Tests/tests/nightwatch/dashboard/dns_mobile.spec.js
@@ -1,8 +1,0 @@
-const tests = require("./dns.spec")
-
-module.exports = {
-    "Resize window": async function (browser) {
-        await browser.resizeWindow(480, 800)
-    },
-    ...tests
-}
